User should be able to create a new fashion item
